refactor: src/model, src/strings
