+++ src/ast/replace-heading.ts
@@ -1,14 +1,15 @@
 import { Heading } from "npm:@types/mdast";
-import { groups, startsWithA } from "../regex.ts";
+import { groups } from "../strings/regex.ts";
 import {
   BOX_REGEX,
   createBoxAndTaskIdPlaceholderRegex,
   createBoxAndTaskIdRegex,
-} from "../strings/box.ts";
-import { ProjectId } from "../strings/project-id.ts";
-import { NextIdentifierNumberGetter } from "../strings/task-id-number.ts";
-import { createTaskIdPlaceholderRegex } from "../strings/task-id-placeholder.ts";
-import { createTaskIdRegex } from "../strings/task-id.ts";
+} from "../model/box.ts";
+import { ProjectId } from "../model/project-id.ts";
+import { NextIdentifierNumberGetter } from "../model/task-id-number.ts";
+import { createTaskIdPlaceholderRegex } from "../model/task-id-placeholder.ts";
+import { createTaskIdRegex } from "../model/task-id.ts";
+import { startsWithA } from "../strings/text-type-guard.ts";
 import { replaceFirstChildTextValue } from "./replace-first-child-text-value.ts";
 import { isWithFirstChildText } from "./types.ts";
