+++ src/ast/replace-parent.ts
@@ -1,7 +1,7 @@
 import { Parent } from "npm:@types/mdast";
 
-import { ProjectId } from "../strings/project-id.ts";
-import { NextIdentifierNumberGetter } from "../strings/task-id-number.ts";
+import { ProjectId } from "../model/project-id.ts";
+import { NextIdentifierNumberGetter } from "../model/task-id-number.ts";
 import { transformNode } from "./transform-node.ts";
 
 export function replaceParent<
