--- src/commands/README.md
@@ -1,5 +1,0 @@
-# src/commands
-
-The commands that we generate as output.
-
-These can be unit-tested without writing to the filesystem.
