+++ src/model/README.md
@@ -1,0 +1,3 @@
+# src/model
+
+Types and simple functions directly related to the model / domain objects.
