+++ src/model/has-protocol.ts
