+++ src/model/inputs.ts
