+++ src/model/is-fragment.ts
