+++ src/model/task-id.ts
@@ -1,12 +1,11 @@
-import { PROJECT_ID_REGEX, type ProjectId } from "./project-id.ts";
+import { extractA } from "../strings/extract-a.ts";
 import {
-  capture,
   containsA,
-  extractA,
-  isA,
-  sequence,
-  TypeGuard,
-} from "../regex.ts";
+  isOnlyA,
+  TextTypeGuard,
+} from "../strings/text-type-guard.ts";
+import { PROJECT_ID_REGEX, type ProjectId } from "./project-id.ts";
+import { capture, sequence } from "../strings/regex.ts";
 import { TASK_ID_NUMBER_REGEX, TaskIdNumber } from "./task-id-number.ts";
 
 export function createTaskIdRegex<PI extends ProjectId = ProjectId>(
@@ -25,8 +24,8 @@ export type TaskId<PI extends ProjectId = ProjectId> = `${PI}-${TaskIdNumber}`;
 
 export function createIsTaskId<PI extends ProjectId = ProjectId>(
   projectId: PI | RegExp = PROJECT_ID_REGEX,
-): TypeGuard<TaskId<PI>> {
-  return isA<TaskId<PI>>(createTaskIdRegex(projectId));
+): TextTypeGuard<TaskId<PI>> {
+  return isOnlyA<TaskId<PI>>(createTaskIdRegex(projectId));
 }
 
 export function createContainsTaskId<PI extends ProjectId = ProjectId>(
